Rename max hit dice attr for players.
The existing attr name conflicted with the NPC sheet, where the value is
static rather than autocalculated. Due to the conflict, the player
formula was unfetchable via the API or chat messages. Renaming the attr
fixes this.

I initially tried renaming the NPC attr, but since that attr is static
and filled in by hand, renaming it would break pre-existing character
sheet attributes. The player attribute on the other hand is
autocalcuated, and since it was unfetchable due to the conflict it
should be unused in any existing macros or API scripts. As such, the
player attribute should be safe to rename.